+++ vss_cli/plugins/compute_plugins/vm.py
@@ -7,7 +7,11 @@ from typing import Optional
 
 import click
 from click_plugins import with_plugins
-from importlib_metadata import entry_points
+
+try:
+    import importlib_metadata as ilm
+except ImportError:
+    import importlib.metadata as ilm
 
 import vss_cli.autocompletion as autocompletion
 from vss_cli import const
@@ -29,7 +33,7 @@ from vss_cli.validators import (
 _LOGGING = logging.getLogger(__name__)
 
 
-@with_plugins(entry_points(group='vss_cli.contrib.compute.vm'))
+@with_plugins(ilm.entry_points(group='vss_cli.contrib.compute.vm'))
 @cli.group('vm', short_help='Manage virtual machines')
 @pass_context
 def compute_vm(ctx: Configuration):