+++ vss_cli/plugins/plugins.py
@@ -3,7 +3,11 @@ import logging
 
 import click
 from click_plugins import with_plugins
-from importlib_metadata import entry_points
+
+try:
+    import importlib_metadata as ilm
+except ImportError:
+    import importlib.metadata as ilm
 
 from vss_cli.cli import pass_context
 from vss_cli.config import Configuration
@@ -11,7 +15,7 @@ from vss_cli.config import Configuration
 _LOGGING = logging.getLogger(__name__)
 
 
-@with_plugins(entry_points(group='vss_cli.contrib.plugins'))
+@with_plugins(ilm.entry_points(group='vss_cli.contrib.plugins'))
 @click.group('plugins', short_help='External plugins.')
 @pass_context
 def cli(ctx: Configuration):